requirements.txt
evdev
pyserial
rplidarc1